+++ pysweep/core/solver.py
@@ -2,8 +2,8 @@ import sys, os, yaml, numpy, h5py, warnings, subprocess, traceback,  mpi4py.MPI 
 import pysweep.core.GPUtil as GPUtil
 import pysweep.core.io as io
 import pysweep.core.process as process
-import pysweep.core.engine as engine
 import pysweep.core.functions as functions
+import pysweep.core.block as block
 
 class Solver(object):
     """docstring for Solver."""
@@ -11,7 +11,7 @@ class Solver(object):
     def __init__(self, initialConditions, yamlFileName=None,sendWarning=True):
         super(Solver, self).__init__()
         self.initialConditions = initialConditions
-        self.arrayshape = numpy.shape(initialConditions)
+        self.arrayShape = numpy.shape(initialConditions)
         self.corepath = os.path.dirname(os.path.abspath(__file__))
         self.libpath = os.path.join(self.corepath,"lib")
         self.kernelpath = os.path.join(self.corepath,"kernels")
@@ -23,7 +23,7 @@ class Solver(object):
             self.initialConditions = self.initialConditions.astype(self.dtype)
 
             #Creating time step data
-            self.createTimeSteps()
+            self.createTimeStepData()
             #set up MPI
             process.setupMPI(self)
             io.verbosePrint(self,"Setting up MPI...\n")
@@ -31,6 +31,12 @@ class Solver(object):
             #Creating simulatneous input and output file
             # io.verbosePrint(self,'Creating output file...\n')
             # self.createOutputFile()
+
+            #Creating shared array
+            if self.simulation:
+                block.sweptBlock(self)
+            else:
+                block.standardBlock(self)
 
         #     #Setting egine
         #     self.engine = "swept" if self.simulation else "standard"
@@ -76,7 +82,7 @@ class Solver(object):
         returnString+="\tcpu source: {}\n".format(self.source['cpu'])
         returnString+="\toutput file: {}\n".format(self.output)
         returnString+="\tverbose: {}\n".format(self.verbose)
-        returnString+="\tArray shape: {}\n".format(self.arrayshape)
+        returnString+="\tArray shape: {}\n".format(self.arrayShape)
         returnString+="\tdtype: {}\n".format(self.dtype)
         returnString+="\tshare: {}\n".format(self.share)
         returnString+="\tblocksize: {}\n".format(self.blocksize[0])
@@ -97,14 +103,17 @@ class Solver(object):
             returnString+="]\n"
         return returnString
 
-    def createTimeSteps(self):
+    def createTimeStepData(self):
         """Use this function to create timestep data."""
         self.time_steps = int((self.timeTuple[1]-self.timeTuple[0])/self.timeTuple[2])  #Number of time steps
         if self.simulation:
-            self.maxPyramidSize = int(self.blocksize[0]/(2*self.operating)-1) #This will need to be adjusted for differing x and y block lengths
+            self.splitx = self.blocksize[0]//2
+            self.splity = self.blocksize[1]//2
+            self.maxPyramidSize = self.blocksize[0]//(2*self.operating)-1 #This will need to be adjusted for differing x and y block lengths
             self.maxGlobalSweptStep = int(self.intermediate*(self.time_steps-self.maxPyramidSize)/(self.maxPyramidSize)+1)  #Global swept step  #THIS ASSUMES THAT time_steps > MOSS
             self.time_steps = int(self.maxPyramidSize*(self.maxGlobalSweptStep+1)/self.intermediate+1) #Number of time
-        self.arrayshape = (self.time_steps,)+self.arrayshape
+            self.maxOctSize = 2*self.maxPyramidSize
+        self.arrayShape = (self.maxOctSize+self.intermediate+1,)+self.arrayShape
 
 
     def createOutputFile(self):
@@ -116,7 +125,7 @@ class Solver(object):
         hdf5_file.create_dataset("exid",(len(self.exid),),data=self.exid)
         hdf5_file.create_dataset("globals",(len(self.globals),),data=self.globals)
         hdf5_file.create_dataset("clocktime",(1,),data=0.0)
-        hdf5_data_set = hdf5_file.create_dataset("data",(self.time_steps,)+self.arrayshape,dtype=self.dtype)
+        hdf5_data_set = hdf5_file.create_dataset("data",(self.time_steps,)+self.arrayShape,dtype=self.dtype)
 
         if self.rank == self.master_rank:
             hdf5_data_set[0,:,:,:] = self.initialConditions[:,:,:]
@@ -188,3 +197,44 @@ class Solver(object):
                 warnings.warn('Code was not recompiled, set recompile=True to force recompile')
                 if not os.path.exists(self.libname):
                     raise UserWarning('{} could not be located'.format(self.libname))
+
+    def standardSolve():
+        # -------------------------------Standard Decomposition---------------------------------------------#
+        node_comm.Barrier()
+        cwt = 1
+        for i in range(TSO*time_steps):
+            functions.Decomposition(GRB,OPS,sarr,garr,blocks,mpi_pool,DecompObj)
+            node_comm.Barrier()
+            #Write data and copy down a step
+            if (i+1)%TSO==0 and NMB:
+                hdf5_data[cwt,i1,i2,i3] = sarr[TSO,:,OPS:-OPS,OPS:-OPS]
+                sarr = numpy.roll(sarr,TSO,axis=0) #Copy down
+                cwt+=1
+            elif NMB:
+                sarr = numpy.roll(sarr,TSO,axis=0) #Copy down
+            node_comm.Barrier()
+            #Communicate
+            functions.send_edges(sarr,NMB,GRB,node_comm,cluster_comm,comranks,OPS,garr,DecompObj)
+
+    def sweptSolve():
+        # -------------------------------SWEPT RULE---------------------------------------------#
+        # -------------------------------FIRST PRISM AND COMMUNICATION-------------------------------------------#
+        functions.FirstPrism(SM,GRB,Up,Yb,mpiPool,blocks,sarr,garr,total_cpu_block)
+        node_comm.Barrier()
+        functions.first_forward(NMB,GRB,node_comm,cluster_comm,comranks,sarr,SPLITX,total_cpu_block)
+        #Loop variables
+        cwt = 1 #Current write time
+        gts = 0 #Initialization of global time step
+        del Up #Deleting Up object after FirstPrism
+        #-------------------------------SWEPT LOOP--------------------------------------------#
+        step = cycle([functions.send_backward,functions.send_forward])
+        for i in range(MGST):
+            functions.UpPrism(GRB,Xb,Yb,Oct,mpiPool,blocks,sarr,garr,total_cpu_block)
+            node_comm.Barrier()
+            cwt = next(step)(cwt,sarr,hdf5_data,gsc,NMB,GRB,node_comm,cluster_comm,comranks,SPLITX,gts,TSO,MPSS,total_cpu_block)
+            gts+=MPSS
+        #Do LastPrism Here then Write all of the remaining data
+        Down.gts = Oct.gts
+        functions.LastPrism(GRB,Xb,Down,mpiPool,blocks,sarr,garr,total_cpu_block)
+        node_comm.Barrier()
+        next(step)(cwt,sarr,hdf5_data,gsc,NMB,GRB,node_comm,cluster_comm,comranks,SPLITX,gts,TSO,MPSS,total_cpu_block)
